+++ external/abl_link_instance.hpp
@@ -39,7 +39,7 @@ class AblLinkWrapper {
   ableton::link::HostTimeFilter<ableton::link::platform::Clock> time_filter;
   std::chrono::microseconds latency_offset;
   t_symbol *num_peers_sym;
-  int num_peers;
+  size_t num_peers;
   double sample_time;
   int invocation_count;
   std::chrono::microseconds curr_time;
